requirements.txt
anyascii==0.3.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
django-filter==23.1
django-ip-restriction==1.1.1
django-ipware==5.0.0
django-phonenumber-field==7.0.2
django-upgrade==1.13.0
idna==3.4
phonenumbers==8.13.9
Pillow==9.5.0
requests==2.28.2
sqlparse==0.4.3
tokenize-rt==5.0.0
tzdata==2023.3
urllib3==1.26.15